functions/requirements.txt
google-cloud-logging
google-cloud-pubsub
pydantic
